requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.2
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.16.0
azure-search-documents==11.6.0b1
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.7
distlib==0.3.1
distro==1.9.0
Django==3.1.4
exceptiongroup==1.2.1
filelock==3.0.12
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
googlesearch-python==1.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
identity==0.8.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
isodate==0.6.1
isort==5.8.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.18
langsmith==0.1.75
lazy-object-proxy==1.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mccabe==0.6.1
mdurl==0.1.2
msal==1.28.0
msal-extensions==1.1.0
msgspec==0.18.6
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.33.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pinecone-plugin-interface==0.0.7
pipenv==2021.5.29
plotly==5.22.0
portalocker==2.8.2
protobuf==4.25.3
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pylint==2.7.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests2==2.16.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.1
streamlit==1.34.0
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
transformers==4.41.2
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.2.2
virtualenv-clone==0.5.6
watchdog==4.0.0
Werkzeug==3.0.3
wordcloud==1.9.3
wrapt==1.12.1
yarl==1.9.4
zipp==3.18.2
